Update v1.0.3
bugs fixes

+++ Packages/player-core-fps/Runtime/MouseSensivity.cs
@@ -2,7 +2,7 @@ namespace InputNS
 {
 	public static class MouseSensivity
 	{
-		public static int XSensitivity { get; set; } = 30;
-		public static int YSensitivity { get; set; } = 30;
+		public static float XSensitivity { get; set; } = 30;
+		public static float YSensitivity { get; set; } = 30;
 	}
 }
